+++ jaseci_core/jaseci/actions/tests/std_test_code.py
@@ -1,0 +1,10 @@
+rand_std = \
+    """
+    walker init
+    {
+        report rand.word();
+        report rand.sentence();
+        report rand.paragraph();
+        report rand.text();
+    }
+    """